moved prof sign out button into the nav bar
