+++ src/components/SignOutBtn.css
@@ -1,0 +1,3 @@
+#log-out-btn {
+	margin-top: 0px !important;
+}
