+++ src/components/pages/Home.js
@@ -1,7 +1,6 @@
 import React from "react";
 import "../../App.css";
-import HeroSection from "../HeroSection";
-import Cards from "../Cards";
+import { Cards, HeroSection } from "../index";
 
 function Home() {
 	return (
